// frontend-nextjs/types/api.ts
export interface ApiResponse<T> {
  data: T;
  message?: string;
  status?: number;
}

export interface ApiError {
  detail: string | Record<string, string[]> | string[];
  msg?: string;
  message?: string;
  status?: number;
}